+++ CHANGES.md
@@ -5,7 +5,7 @@
 ##### Additions :tada:
 
 - Added `Matrix3.inverseTranspose` and `Matrix4.inverseTranspose`. [#9135](https://github.com/CesiumGS/cesium/pull/9135)
-- Added `Camera.maxTilt` to prevent camera from being tilted past that angle.
+- Added `Camera.maximumTilt` to prevent camera from being tilted past that angle.
 
 ##### Fixes :wrench: